+++ src/utils/setup/validate-env-variables.js
@@ -1,0 +1,17 @@
+// env variables for building site along with use in front-end
+// https://www.gatsbyjs.org/docs/environment-variables/#defining-environment-variables
+const validateEnvVariables = () => {
+  // make sure necessary env vars exist
+  if (!process.env.GATSBY_SITE || !process.env.PARSER_USER || !process.env.PARSER_BRANCH) {
+    return {
+      error: true,
+      message: `${process.env.NODE_ENV} requires the variables GATSBY_SITE, PARSER_USER, and PARSER_BRANCH`,
+    };
+  }
+  // create split prefix for use in stitch function
+  return {
+    error: false,
+  };
+};
+
+module.exports.validateEnvVariables = validateEnvVariables;
